split command struct out into own file
